+++ tests/ResponseBuilderTest.php
@@ -129,6 +129,25 @@ class ResponseBuilderTest extends TestCase
     /**
      * @test
      */
+    public function itCanBuildAResponseWithIgnoredParameters(): void
+    {
+        // arrange
+        $messageFactory = MessageFactoryDiscovery::find();
+        $builder = $this->getBuilder()->setIgnoredQueryParameters(['ignore-me', 'ignore-me-too']);
+
+        $expectedResponse = $messageFactory->createResponse()
+            ->withBody(Utils::streamFor(file_get_contents($this->getFixturesPath().'/example.com/api/articles.mock')));
+
+        // act
+        $actualResponse = $builder->build($messageFactory->createRequest('GET', 'https://example.com/api/articles?ignore-me=true&ignore-me-too'));
+
+        // assert
+        $this->assertEquals($expectedResponse->getBody()->__toString(), $actualResponse->getBody()->__toString());
+    }
+
+    /**
+     * @test
+     */
     public function itCanBuildAResponseWithCustomHeaders()
     {
         $messageFactory = MessageFactoryDiscovery::find();